requirements.txt
Flask==1.1.2
google-api-python-client==1.10.0
google-auth==1.19.2
google-cloud-asset==1.3.0
requests==2.24.0
